+++ blaze-client/src/test/scala/org/http4s/client/blaze/Http1ClientStageSpec.scala
@@ -5,8 +5,8 @@ package blaze
 import java.nio.charset.StandardCharsets
 import java.nio.ByteBuffer
 
-import org.http4s.blaze.SeqTestHead
 import org.http4s.blaze.pipeline.LeafBuilder
+import org.http4s.blazecore.SeqTestHead
 import org.http4s.util.threads.DefaultPool
 import bits.DefaultUserAgent
 import org.specs2.mutable.Specification
