+++ blaze-core/src/main/scala/org/http4s/blazecore/Http1Stage.scala
@@ -1,5 +1,5 @@
 package org.http4s
-package blaze
+package blazecore
 
 import java.nio.ByteBuffer
 import java.nio.charset.StandardCharsets
@@ -11,6 +11,7 @@ import org.http4s.blaze.util.BufferTools.{concatBuffers, emptyBuffer}
 import org.http4s.blaze.http.http_parser.BaseExceptions.ParserException
 import org.http4s.blaze.pipeline.{Command, TailStage}
 import org.http4s.blaze.util._
+import org.http4s.blazecore.util._
 import org.http4s.util.{Writer, StringWriter}
 import scodec.bits.ByteVector
