+++ blaze-core/src/main/scala/org/http4s/blazecore/util/BodylessWriter.scala
@@ -1,5 +1,5 @@
 package org.http4s
-package blaze
+package blazecore
 package util
 
 import java.nio.ByteBuffer
