+++ blaze-core/src/main/scala/org/http4s/blazecore/util/CachingChunkWriter.scala
@@ -1,4 +1,6 @@
-package org.http4s.blaze.util
+package org.http4s
+package blazecore
+package util
 
 import java.nio.ByteBuffer
