+++ blaze-core/src/main/scala/org/http4s/blazecore/util/Http2Writer.scala
@@ -1,5 +1,6 @@
-package org.http4s.blaze.util
-
+package org.http4s
+package blazecore
+package util
 
 import org.http4s.blaze.http.Headers
 import org.http4s.blaze.pipeline.TailStage
