+++ blaze-core/src/main/scala/org/http4s/blazecore/util/ProcessWriter.scala
@@ -1,5 +1,5 @@
 package org.http4s
-package blaze
+package blazecore
 package util
 
 import scodec.bits.ByteVector
