+++ blaze-core/src/test/scala/org/http4s/blazecore/TestHead.scala
@@ -1,8 +1,9 @@
-package org.http4s.blaze
+package org.http4s
+package blazecore
 
 import org.http4s.blaze.pipeline.HeadStage
 import org.http4s.blaze.pipeline.Command.{Disconnected, Disconnect, OutboundCommand, EOF}
-
+import org.http4s.blaze.util.TickWheelExecutor
 import java.nio.ByteBuffer
 
 import scala.concurrent.duration.Duration
@@ -11,6 +12,7 @@ import scala.util.{Success, Failure, Try}
 
 
 abstract class TestHead(val name: String) extends HeadStage[ByteBuffer] {
+  val scheduler = new TickWheelExecutor()
 
   private var acc = Vector[Array[Byte]]()
   private val p = Promise[ByteBuffer]
@@ -33,6 +35,7 @@ abstract class TestHead(val name: String) extends HeadStage[ByteBuffer] {
 
   override def stageShutdown(): Unit = synchronized {
     closed = true
+    scheduler.shutdown()
     super.stageShutdown()
     p.trySuccess(ByteBuffer.wrap(getBytes()))
     ()
@@ -53,7 +56,6 @@ class SeqTestHead(body: Seq[ByteBuffer]) extends TestHead("SeqTestHead") {
 }
 
 final class SlowTestHead(body: Seq[ByteBuffer], pause: Duration) extends TestHead("Slow TestHead") { self =>
-  import org.http4s.blaze.util.Execution.scheduler
 
   private val bodyIt = body.iterator
   private var currentRequest: Option[Promise[ByteBuffer]] = None
