+++ blaze-core/src/test/scala/org/http4s/blazecore/util/FailingWriter.scala
@@ -1,4 +1,6 @@
-package org.http4s.blaze.util
+package org.http4s
+package blazecore
+package util
 
 import org.http4s.blaze.pipeline.Command.EOF
 import scodec.bits.ByteVector
