+++ blaze-server/src/main/scala/org/http4s/server/blaze/Http1ServerStage.scala
@@ -3,13 +3,13 @@ package server
 package blaze
 
 
+import org.http4s.blaze.http.http_parser.BaseExceptions.{BadRequest, ParserException}
 import org.http4s.blaze.pipeline.Command.EOF
-import org.http4s.blaze.Http1Stage
 import org.http4s.blaze.pipeline.{Command => Cmd, TailStage}
-import org.http4s.blaze.util.BodylessWriter
 import org.http4s.blaze.util.Execution._
 import org.http4s.blaze.util.BufferTools.emptyBuffer
-import org.http4s.blaze.http.http_parser.BaseExceptions.{BadRequest, ParserException}
+import org.http4s.blazecore.Http1Stage
+import org.http4s.blazecore.util.BodylessWriter
 
 import org.http4s.internal.compatibility._
 import org.http4s.util.StringWriter
